+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.0] - 2026-07-11
+- Added `client` option to identify integrations of the l10n.dev localization platform.
+
 ## [1.9.1] - 2026-07-06
 - Updated README to refer local MCP server