+++ sdk/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the SDK package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.0] - 2026-07-11
+- Added `authOptions` as an optional arg to the `translate` method of `AITranslator` class.
+  - Add `client` to identify client's usage correctly
+  - Moved `apiKey` to the `authOptions`.
+
+### Breaking changes
+ - Optional `apiKey` arg now in the `authOptions`
+
 ## [1.9.1] - 2026-07-06
 - Added BalanceManager. Updated README to refer local MCP server